backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.1.0
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
